Move iframe logic to page component
