+++ src/app/layout/embed-layout/embed-layout-component.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import MapIframe from 'pages/map-iframe';
+
+const iFrameLayoutComponent = () => {
+  return <MapIframe />;
+};
+
+export default iFrameLayoutComponent;
