+++ src/app/layout/embed-layout/embed-layout-styles.scss
@@ -1,0 +1,1 @@
+@import '~styles/variables.scss';
