+++ src/app/layout/embed-layout/embed-layout.js
@@ -1,0 +1,1 @@
+export { default } from './embed-layout-component';
